Feat/12/ensure test coverage

+++ src/controllers/CartController.js
@@ -3,15 +3,9 @@ const { Cart_Item, Shopping_Session } = require('../models')
 const { QueryTypes } = require('sequelize')
 const models = require('../models')
 
-const UsuarioEnderecosController = {
+const CartController = {
 
   show: async (req, res) => {
-    const { user } = req.session
-
-    // caso o usuario nao esteja logado, redirecionar para a pagina de login
-    if (!user) {
-      res.redirect('/login')
-    }
     const { shopping_session } = req.session.user
     try {
       const cartDetails = await Cart_Item.findAll({
@@ -20,8 +14,9 @@ const UsuarioEnderecosController = {
         },
         include: ['product', 'shoppingsession']
       })
+      // console.log('cartDetails', cartDetails)
 
-      if (cartDetails.length === 0) {
+      if (cartDetails?.length === 0) {
         return res.status(200).render('carrinho', {
           arquivoCss: 'carrinho.css',
           error: 'Não há nenhuma informação cadastrada.',
@@ -34,7 +29,7 @@ const UsuarioEnderecosController = {
         cartItens: cartDetails
       })
     } catch (error) {
-      return res.status(500).json({ message: 'Error' + error })
+      return res.status(500).render({ message: 'Error' + error })
     }
   },
   addProduct: async (req, res) => {
@@ -49,12 +44,12 @@ const UsuarioEnderecosController = {
         include: ['product', 'shoppingsession']
       })
 
-      // if (!cartItem) {
-      //   return res.status(422).render('carrinho', {
-      //     arquivoCss: 'carrinho.css',
-      //     error: 'Pedido não localizado!'
-      //   })
-      // }
+      if (!cartItem) {
+        return res.status(422).render('carrinho', {
+          arquivoCss: 'carrinho.css',
+          error: 'Pedido não localizado!'
+        })
+      }
 
       const newItemData = {
         quantity: cartItem.quantity + 1,
@@ -73,6 +68,7 @@ const UsuarioEnderecosController = {
 
       // update Shopping_Session with new total value
       await Shopping_Session.update({
+        individualHooks: true,
         total: cartTotalPrice[0].newTotal,
         updatedAt: new Date().toISOString()
       }, {
@@ -83,7 +79,7 @@ const UsuarioEnderecosController = {
 
       return res.redirect('/carrinho')
     } catch (error) {
-      return res.status(500).json({ message: 'Error' + error })
+      return res.status(500).render({ message: 'Error' + error })
     }
   },
   delProduct: async (req, res) => {
@@ -98,37 +94,39 @@ const UsuarioEnderecosController = {
         include: ['product', 'shoppingsession']
       })
 
-      // if (!cartItem) {
-      //   return res.status(422).render('carrinho', {
-      //     arquivoCss: 'carrinho.css',
-      //     error: 'Pedido não localizado!'
-      //   })
-      // }
+      if (!cartItem) {
+        return res.status(422).render('carrinho', {
+          arquivoCss: 'carrinho.css',
+          error: 'Pedido não localizado!'
+        })
+      }
 
+      
       const newItemData = {
         quantity: cartItem.quantity - 1,
         updatedAt: new Date().toISOString()
       }
 
       await Cart_Item.update(newItemData, {
+        individualHooks: true,
         where: {
           id: cartItem.id
         }
       })
-
       // calculate total value of Cart Itens
-      const cartTotalPrice = await models.sequelize.query(`SELECT SUM(p.price * ci.quantity) as newTotal FROM go_pedal.Product p , go_pedal.Cart_Item ci  ,go_pedal.Shopping_Session ss  WHERE ss.id = ${ShoppingSessionId} AND ss.id = ci.ShoppingSessionId AND ci.ProductId =p.id`, { type: QueryTypes.SELECT })
+      const cartTotalPrice = await models.sequelize.query(`SELECT SUM(p.price * ci.quantity) as newTotal FROM Product p , Cart_Item ci, Shopping_Session ss WHERE ss.id = ${ShoppingSessionId} AND ss.id = ci.ShoppingSessionId AND ci.ProductId =p.id`, { type: QueryTypes.SELECT })
 
-      if (cartTotalPrice[0].newTotal < 0) {
-        return res.status(422).render('carrinho', {
-          arquivoCss: 'carrinho.css',
-          cartItens: {},
-          error: 'O produto não pode ter valor negativo. Exclua-o primeiro.'
-        })
-      }
+      // if (cartTotalPrice[0].newTotal < 0) {
+      //   return res.status(422).render('carrinho', {
+      //     arquivoCss: 'carrinho.css',
+      //     cartItens: {},
+      //     error: 'O produto não pode ter valor negativo. Exclua-o primeiro.'
+      //   })
+      // }
 
       // update Shopping_Session with new total value
       await Shopping_Session.update({
+        individualHooks: true,
         total: cartTotalPrice[0].newTotal,
         updatedAt: new Date().toISOString()
       }, {
@@ -139,7 +137,7 @@ const UsuarioEnderecosController = {
 
       return res.redirect('/carrinho')
     } catch (error) {
-      return res.status(500).json({ message: 'Error' + error })
+      return res.status(500).render({ message: 'Error' + error })
     }
   },
   removeProduct: async (req, res) => {
@@ -154,12 +152,12 @@ const UsuarioEnderecosController = {
         include: ['product', 'shoppingsession']
       })
 
-      // if (!cartItem) {
-      //   return res.status(422).render('carrinho', {
-      //     arquivoCss: 'carrinho.css',
-      //     error: 'Pedido não localizado!'
-      //   })
-      // }
+      if (!cartItem) {
+        return res.status(422).render('carrinho', {
+          arquivoCss: 'carrinho.css',
+          error: 'Pedido não localizado!'
+        })
+      }
 
       await Cart_Item.destroy({
         where: {
@@ -171,9 +169,9 @@ const UsuarioEnderecosController = {
             // calculate total value of Cart Itens
             const cartTotalPrice = await models.sequelize.query(`SELECT SUM(p.price * ci.quantity) as newTotal FROM go_pedal.Product p , go_pedal.Cart_Item ci  ,go_pedal.Shopping_Session ss  WHERE ss.id = ${ShoppingSessionId} AND ss.id = ci.ShoppingSessionId AND ci.ProductId =p.id`, { type: QueryTypes.SELECT })
 
-            if (!cartTotalPrice[0].newTotal) {
-              cartTotalPrice[0].newTotal = 0
-            }
+            // if (!cartTotalPrice[0].newTotal) {
+            //   cartTotalPrice[0].newTotal = 0
+            // }
 
             // update Shopping_Session with new total value
             await Shopping_Session.update({
@@ -185,63 +183,49 @@ const UsuarioEnderecosController = {
               }
             })
             return res.redirect('/carrinho')
-          } else {
-            return res.status(404).render('404', {
-              textoErro: 'Categoria não encontrado, refaça sua busca ou tente novamente'
-            })
           }
         })
     } catch (error) {
-      return res.status(500).json({ message: 'Error' + error })
+      return res.status(500).render({ message: 'Error' + error })
     }
   },
   includeProduct: async (req, res) => {
     try {
-      const { user } = req.session
-
-      // caso o usuario nao esteja logado, redirecionar para a pagina de login
-      if (!user) {
-        res.redirect('/login')
-      }
-
       const ProductId = req.params.id
       const ShoppingSessionId = req.session.user.shopping_session
 
-      await Cart_Item.findOrCreate({
-        where: {
-          ProductId,
-          ShoppingSessionId
-        },
-        defaults: {
-          quantity: 1
-        }
-      })
-
-      // if (!cartItem) {
-      //   return res.status(422).render('carrinho', {
-      //     arquivoCss: 'carrinho.css',
-      //     error: 'Pedido não localizado!'
-      //   })
-      // }
+      try {
+        await Cart_Item.findOrCreate({
+          where: {
+            ProductId,
+            ShoppingSessionId
+          },
+          defaults: {
+            quantity: 1
+          }
+        })
+      } catch (error) {
+        return res.status(500).render({ message: 'Error' + error })
+      }
 
       // calculate total value of Cart Itens
-      const cartTotalPrice = await models.sequelize.query(`SELECT SUM(p.price * ci.quantity) as newTotal FROM go_pedal.Product p , go_pedal.Cart_Item ci  ,go_pedal.Shopping_Session ss  WHERE ss.id = ${ShoppingSessionId} AND ss.id = ci.ShoppingSessionId AND ci.ProductId =p.id`, { type: QueryTypes.SELECT })
+      const cartTotalPrice = await models.sequelize.query(`SELECT SUM(p.price * ci.quantity) as newTotal FROM go_pedal.Product p , go_pedal.Cart_Item ci  ,go_pedal.Shopping_Session ss  WHERE ss.id = ${ShoppingSessionId} AND ss.id = ci.ShoppingSessionId AND ci.ProductId = p.id`, { type: QueryTypes.SELECT })
 
       // update Shopping_Session with new total value
-      await Shopping_Session.update({
-        total: cartTotalPrice[0].newTotal,
-        updatedAt: new Date().toISOString()
-      }, {
-        where: {
-          id: ShoppingSessionId
-        }
-      })
+        await Shopping_Session.update({
+          total: cartTotalPrice[0].newTotal,
+          updatedAt: new Date().toISOString()
+        }, {
+          where: {
+            id: ShoppingSessionId
+          }
+        })
 
       return res.redirect('/carrinho')
     } catch (error) {
-      return res.status(500).json({ message: 'Error' + error })
+      return res.status(500).render({ message: 'Error' + error })
     }
   }
 }
 
-module.exports = UsuarioEnderecosController
+module.exports = CartController
